% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diveristy.R
\name{diveristy}
\alias{diveristy}
\title{Genetic diversity}
\usage{
diveristy(data)
}
\arguments{
\item{data}{A genotype frame of SNPs in HMP formate}
}
\value{
A data frame about genetic diversity that can discribe the SNPs.
}
\description{
You can use "diveristy()" to compute the genetic diversity of SNPs.
}
\examples{
data("E5b_hmp")
diveristy(E5b_hmp)
}
