% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GenoToHmp.R
\name{GenoToHmp}
\alias{GenoToHmp}
\title{Convert the GENO format to HMP format}
\usage{
GenoToHmp(tsv, genename)
}
\arguments{
\item{tsv}{the geno format of SNPs (that the output format of novoSNP)}

\item{genename}{Genename of SNPs}
}
\value{
A date frame of SNPs in HMP format
}
\description{
Convert the GENO format to HMP format
}
\examples{
data(E5b_geno)
GenoToHmp(E5b_geno, "E5b")
}
