% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ld_plot.R
\name{ld_plot}
\alias{ld_plot}
\title{LD block heatmap}
\usage{
ld_plot(data, pos)
}
\arguments{
\item{data}{A genotype frame of SNPs in HMP format}

\item{pos}{The position of snps on chromosomes}
}
\value{
The file named LDheatmap.pdf will be generated in the current
directory.
}
\description{
ld_plot() is used to produce a graphical display, as a heat map, of pairwise
linkage disequilibrium (LD) measurements for SNPs.
}
\examples{
library(genetic) # 加载包
data("E5b_hmp") # 读取数据
ld_plot(E5b_hmp, E5b_hmp$pos) # 在当前文件夹下生成LDheatmap.pdf
}
